backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
werkzeug==3.0.3
scikit-learn==1.8.0
numpy==2.1.3
pandas==2.2.2
python-dotenv==1.0.1
bcrypt==4.1.3
reportlab==4.2.2
